Style change: remove type encoding from all var/member/parameter names
Explanation: I wasn't really a fan of this from the beginning, the habit
slipped in because of the company I currently work for. But actually I
agree with the C++ language inventors on this topic. Andsince there are
no collaborators on this project (yet) and the code base is still of
manageable size I decided to do this change now (actually I used a few
regexes to accomplish this because the code base is already too big to
do this manually ;) ).

See https://en.wikipedia.org/wiki/Hungarian_notation#Notable_opinions