NoteRgbSource does not create a default RGB function
I did not like the dependency from the first to the latter. Also, the
unit test should not be testing LinearRgbFunction, that one should get
its own unit test instead.

The unit test is fixed by setting a mock RGB function (as it should be)
and setting an action with a result similar to the former default
implementation.

Also tidied up some includes (#18) and did some style fixes (#22)